+++ mitra_activitypub/src/actors/handlers.rs
@@ -8,8 +8,8 @@ use uuid::Uuid;
 
 use apx_core::{
     http_url::Hostname,
+    http_url_whatwg::get_hostname,
     url::canonical::CanonicalUrl,
-    urls::get_hostname,
 };
 use apx_sdk::{
     addresses::WebfingerAddress,