+++ mitra_activitypub/src/deliverer.rs
@@ -27,13 +27,13 @@ use apx_core::{
         RsaPublicKey,
         RsaSecretKey,
     },
+    http_url_whatwg::get_hostname,
     json_signatures::create::{
         is_object_signed,
         sign_object,
         JsonSignatureError,
     },
     url::hostname::is_onion,
-    urls::get_hostname,
 };
 use apx_sdk::{
     deliver::{send_object, DelivererError},