+++ mitra_api/src/server.rs
@@ -22,7 +22,7 @@ use actix_web::{
 };
 use log::Level;
 
-use apx_core::urls::get_hostname;
+use apx_core::http_url_whatwg::get_hostname;
 use mitra_config::{Config, Environment};
 use mitra_models::database::DatabaseConnectionPool;
 use mitra_services::{