+++ apx_sdk/CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Return decoded signature string from `parse_http_signature`.
 - Renamed `HttpSignatureData.message` field to `base`.
 - Add header name to `HttpSignatureVerificationError::HeaderError` messages.
+- Renamed `apx_core::urls` module to `http_url_whatwg`.
 
 ### Removed